Example spectra used by Spectran

Contains Measurement and CIE Standard Illuminant Spectra. Additional data to the spectra can be found in the examplespectra_descriptor dataset.

Usage

examplespectra

Format

examplespectra A list of two elements, each containing a table of spectral data

Measurement

Spectral data measured by a spectroradiometer. The unit of the wavelength (Wellenlaenge) is nm, the unit of the spectral data is W/m^2.

CIE

Spectral data from the CIE describing Standard Illuminants. The unit of the wavelength (Wellenlaenge) is nm, the unit of the spectral data is W/m^2.

Source

https://cie.co.at/data-tables